requirements.txt
data==0.4
decorator==5.1.1
funcsigs==1.0.2
glcontext==2.3.7
moderngl==5.8.2
numpy==1.24.3
Pillow==9.5.0
pygame==2.4.0
PyGLM==2.7.0
PyWavefront==1.3.3
six==1.16.0